% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CO2Inhalation-data.R
\docType{data}
\name{CO2Inhalation}
\alias{CO2Inhalation}
\title{CO2 Inhalation Data}
\format{Dataframe with 239 rows and 10 columns.
 The first column indicates the experimental phase and the last nine columns correspond to the nine physiological measures tracked during the experiment:
 Breathing volume variables (ViVol, VeVol, Vent, PiaAB), breathing duration variables (Ti,Te,Tt), heart rate (HR) and RR interval (RR) or cardiac beat interval.}
\usage{
data(CO2Inhalation)
}
\description{
Nine physiological measures during a CO2-inhalation experiment.
}
\examples{
data(CO2Inhalation)
}
\references{
{ \cite{De Roover, K., Timmerman, M. E., Van Diest, I., Onghena, P., & Ceulemans, E. (2014). Switching principal component analysis for modeling means and covariance changes over time. Psychological Methods, 19, 113-132.} \url{https://doi.org/10.1037/a0034525}}
}
\keyword{datasets}
